+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Fixed web minidump upload to work with zero attachments
 - Fixed web crash list from showing all rows as "Processing"
-
+- Fixed symbol lists not showing symbol count for all os's
+- Fixed linux icon not showing in symbol list
 
 ## [0.3.0] - 2021-09-29
 ### Added